+++ source/render/hwgl/gl_draw.cpp
@@ -268,7 +268,7 @@ void VOpenGLDrawer::StartAutomap () {
 //  VOpenGLDrawer::DrawLine
 //
 //==========================================================================
-void VOpenGLDrawer::DrawLine (int x1, int y1, vuint32 c1, int x2, int y2, vuint32 c2) {
+void VOpenGLDrawer::DrawLine (float x1, float y1, vuint32 c1, float x2, float y2, vuint32 c2) {
   guard(VOpenGLDrawer::DrawLine);
   SetColour(c1);
   glVertex2f(x1, y1);
